Do not compute bishop and rook attack-tables at compile time
* What were you thinking...
