+++ src/Sliders.cpp
@@ -1,0 +1,6 @@
+#include "Sliders.hpp"
+
+namespace Brainiac {
+    std::array<SlidingMoveTable, 64> ROOK_ATTACK_TABLES = {};
+    std::array<SlidingMoveTable, 64> BISHOP_ATTACK_TABLES = {};
+} // namespace Brainiac
