+++ tests/CMakeLists.txt
@@ -7,17 +7,10 @@ set(CMAKE_CXX_STANDARD_REQUIRED on)
 
 add_compile_options(-Wall -g -O3 -fPIC)
 
-if ( CMAKE_CXX_COMPILER_ID STREQUAL "GNU" )
-    add_compile_options(-fconstexpr-ops-limit=99999999)
-elseif( CMAKE_CXX_COMPILER_ID MATCHES "^(Apple)?Clang$" )
-    add_compile_options(-fconstexpr-steps=99999999)
-endif()
-
-
 file(GLOB_RECURSE INCLUDE "./src/*.hpp" "../src/*.hpp")
 file(GLOB_RECURSE SOURCES "../src/*.cpp")
 
-list(FILTER SOURCES EXCLUDE REGEX ".*Brainiac\\.cpp$")
+list(FILTER SOURCES EXCLUDE REGEX ".*main\\.cpp$")
 
 # Compile all tests
 file(GLOB TESTS "./src/*.cpp")
