+++ tests/src/Bitboard-test.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "../../src/Bitboard.hpp"
+#include "../../src/Brainiac.hpp"
 
 #include "ctest.hpp"
 
@@ -87,6 +87,7 @@ static char *all_tests() {
 }
 
 int main(int argc, char **argv) {
+    init();
     char *result = all_tests();
     if (result != 0) {
         std::cout << "FAILED... " << result << "\n";
