+++ tests/src/Perft-test.cpp
@@ -1,7 +1,7 @@
 #include <array>
 #include <iostream>
 
-#include "../../src/Perft.hpp"
+#include "../../src/Brainiac.hpp"
 
 #include "ctest.hpp"
 
@@ -97,6 +97,7 @@ static char *all_tests() {
 }
 
 int main(int argc, char **argv) {
+    init();
     char *result = all_tests();
     if (result != 0) {
         std::cout << "FAILED... " << result << "\n";
