+++ tests/src/Utils-test.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "../../src/Utils.hpp"
+#include "../../src/Brainiac.hpp"
 
 #include "ctest.hpp"
 
@@ -33,6 +33,7 @@ static char *all_tests() {
 }
 
 int main(int argc, char **argv) {
+    init();
     char *result = all_tests();
     if (result != 0) {
         std::cout << "FAILED... " << result << "\n";
